Move print_range from Algorithm.test to Panlib.Range.io, add Makefile to Panlib.Range, add range concepts, add range extention helper, fix iterator ranges.
[concept.h]
This provides the way to check which category Range is.
Also this provides various implement helpers that delegate procedures to their child classes, and automatical chooser for these.

[extends.h]
This provides a implement helper for transform range.

[constants.h]
This provides some types for compile-time logical calculations.

[crtp_cast.h]
This provides a cast from a base class to the derived class in the situation of CRTP.